+++ package.json
@@ -12,6 +12,8 @@
     "node-sass": "^4.9.2",
     "prop-types": "^15.6.2",
     "react-fontawesome": "^1.6.1",
+    "react-router": "^4.3.1",
+    "react-router-dom": "^4.3.1",
     "reactstrap": "^6.3.0",
     "sass-loader": "^7.0.3"
   },
