+++ src/components/cart/cart.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faShoppingCart } from '@fortawesome/free-solid-svg-icons';
-import { CartManager } from '../catalog_page.js';
+import { CartManager } from '../app.js';
 import './cart.scss';
 
 export default class Cart extends Component {
