+++ src/components/catalog_page.js
@@ -1,59 +1,15 @@
 import React from 'react';
-import  { bind, find, assign, reduce, includes }  from 'lodash';
 import Catalog from './catalog/catalog.js';
-import CartContent from './cart_content/cart_content.js';
-import Products from '../constants/products.js';
-export const CartManager = React.createContext();
+import Nav from './nav/nav.js';
 
 export default class CatalogPage extends React.Component{
-  constructor(props){
-    super(props)
-    this.state = { 
-      products: Products,
-      cart: {
-       entries: [], 
-       isDragging: false
-      } 
-    }
-    this.addToCart = bind(this.addToCart, this)
-  }
-
-  addToCart(product, quantity=1) {
-    const { entries } = this.state.cart
-    let newEntries =  entries.slice();
-
-    if(includes(newEntries, product)){
-      find(newEntries, product).quantity += quantity
-    }else{
-      newEntries.push(assign(product, { quantity }))
-    }
-
-    this.setState({cart: {entries: newEntries}})
-  }
-
-  getProducts = () => this.state.cart.entries
-  getProduct = (id) => find(this.state.products, { id })
-
-  getTotalCartEntries = () =>  reduce(
-    this.getProducts(), function(acc, item){
-      return acc + item.quantity
-    }, 0)
-
   render() {
-    const { entries } = this.state.cart
-    const { products } = this.state
+    const { products } = this.props
     return (
-      <CartManager.Provider value={
-        {
-          addToCart: this.addToCart, 
-          getProducts: this.getProducts(),
-          getTotalCartEntries: this.getTotalCartEntries(),
-          getProduct: this.getProduct
-        }}
-      >
+      <div>
+        <Nav />
         <Catalog products={ products } />
-        <CartContent />
-      </CartManager.Provider>
+      </div>
     );
   }
 }
